app/flask-app/flask/src/requirements.txt
flask
flask-cors
uwsgi
psycopg2
black
flake8
grpcio
grpcio-tools